Integrated new BlockLabel and fixed snapsegments
In the file integrity section of SegmentManager, the new BlockLabel
function is called. In testing this, I found an error in the
segment replacement routine snapsegments.m. The case is as follows:
for the surface trace of a mesh that involves a triple junction
(along Cascadia, for example), the selected segments had their nodes
modified, but the intersecting segments did not, because they were
not part of the selection. Now, triple junction segments are sought
and the nodes they share with the mesh-replaced segments are modified.
This results in closed blocks.