New BlockLabel is re-implemented; segmentmidpoint.m updated
The issue was not with the code itself, but the fact that, for the
test Japan model that we were running, no stations existing on the
exterior block. One station (in northern Russia) *appeared* to be
outside the Eurasia block, but given the new way that BlockLabel is
run, using spherical coordinates, this station is infact within the
Eurasia block. I will make an issue to add a new check for
SegmentManager to do the formal station labeling, giving the user a
warning if any blocks have no stations on them, since it can be hard
to tell visually that this is the case for high-latitude blocks and
stations.

segmentmidpoint.m was updated to catch north-south segments, which
formerly were assigned a NaN midpoint latitude.