Initial commit of VelCombo
Tools added for combining velocity fields based on alignment of
vectors at stations collocated in the different fields.

This initial version works but isn't super clean. An example run
using the call:

AlignAllFields6Param('./', 2, 2, 6)

seemed to work, aligning all .sta files that appear within the
current directory to the reference frame Field 2 (alphabetically;
I believe the second argument can also be a strng specifying the
name of the file to which you want to align). The third argument
tells the code to use method 2 in constructing the "connection
tree" between fields, which connects one field to the other whose
collocated velocities vary least. Method 1 connects fields based on
the maximum number of shared stations. The 6 at the end tells it to
combined fields that share at least 6 stations (the default is 10).
I'll try to clean this stuff up soon.

Right now, the biograph stuff throws an error/warning, and there's a
keyboard somewhere, but it does write the new combined fields and makes
a not-super-useful global velocity plot when it's done.