SegmentManager block checking updated
- "Check-Ips" function was modifed as "Block Checker". In addition
to checking for interior points, the function also highlights any
block that does not contain any stations. The try/catch structure,
and the structure of BlockLabel in general, means that only when
every block has an interior point will the zero-station blocks be
highlighted (but lack of an interior point is a bigger problem,
anyway, so perhaps this makes sense). A future idea could be to
modify BlockLabel so that it won't throw an error, just a warning,
if not every circulated block has an interior point, but I'm always
hesitant to mess with BlockLabel.

- Most buttons and drop-downs now have a tool tip, so when you hover
the mouse over them, a text window appears giving a brief description
of the GUI element. These can get a little distracting, but I found
that with typical usage of quickly clicking on buttons, it's not bad,
and moving forward it seems like a nice addition for new Blocks users.