GetElasticPartials; variable saving changes for global models
Changed GetElasticPartials to do the segment-local oblique Mercator
within the for loop as opposed to the previous method of doing all
segments and stations as part of a single matrix. This substantially
reduces RAM usage for huge segment files.

Added a few "-v7.3" flags to saving commands so that multiple
gigabyte variables can be saved (elastic partials, for example).