New Blocks command line option and other minor changes
- A new command line option allows Blocks to be called with a
given command file but then have any valid field of the Command
structure specified as an input argument. For example, to run Blocks
with the command file "test.command" but specifying a smoothing weight
of 0.1 rather than whatever is contained in the command file, run
"Blocks('test.command', 'triSmooth', 0.1);" The .command file copied
to the results directory is updated with the altered input arguments.
This can be really handy for running Blocks within a loop, using a
base command file but slightly modifying some of the model parameters
(I've been using it to run time-sequential Blocks models for Japan).
- Changed the distance dependence of triangular smoothing, moving
from the weighting back to the actual derivative approximation (this
was the case for a while back). Overall, this seems to permit estimation
of more detail for a given smoothing value, but I do need to run some
resolution tests.