ResultManager improvements
- New addition to accept directory name(s) on the command line,
which automatically loads results for the specified run. For
example, call ResultManager('0000000001', '0000000002') to avoid
having to use the GUI file picker. A single result directory works
too. Naturally this involves a delay in drawing the figure window,
because the results files are being loaded before drawing.
- Scaling vectors and residual comparison bubbles now works, at
least quantitatively. The legend in the upper left corner giving
the scaling is off, and this is a known bug that I will create an
issue for.
- Strain axis scaling does not work.