Fixed SegmentManager issues with setting segment properties.
1. In selecting a single segment, changed

setappdata(gcf, 'segIdx', segIdx+2);

to

setappdata(gcf, 'segIdx', segIdx);

The +2 is necessary for getting the right name of the highlighted
segment to show up in the name drop-down menu, because of the "blank"
and "Multiple" entries at the top of the name list, but it's not
necessary (and in fact is incorrect) for setting the true index of
the segment.

2. After the selected segment property was changed, the segments are
redrawn, and in doing so, alphabetically sorted. This resulted in a
different segment being highlighted as selected, as segments had
not been sorted upon loading. This behavior is now corrected.