Updated snapsegments and AllTriConstraints
snapsegments.m (within SegmentManager) and AllTriConstraints.m now use
edgeelements.m to find the nodes/elements that line the edges of a mesh.
This is more reliable than the prior method for finding mesh edges,
particularly in the case of irregular depth nodes along the up- and
down-dip edges of a subduction zone mesh.