Updated AllTriConstraints.m to support multiple a priori slip files
There are now two options for specifying a priori triangular slip
or coupling constraints:
1) One file, specified in the .command file
2) One or more files, specified as every 4th line in the .mshp file.
In this case, blank lines should still be retained in the .mshp file
for meshes that do not have any triangular constraints.

For Case 1, specify element indices in terms of the indices within
the range 1:sum(p.nEl). For Case 2, specify element indices local
to the single mesh described by the a priori file.