Fixed edgelements.m to work more reliably with multiple meshes.
When setting triangular edge slip constraints, edgeelements.m finds
the top, bottom, and side edge elements. This works by passing the
entire coordinate array along with an individual mesh's vertex
array to edgeelements.m. The distinction between the top and bottom
edges is made by comparing the edge depth to the mean depth of all
coordinates, but this can return an incorrect identification of a
top or bottom when the entire coordinate array is passed. The simple
fix is to compare a given edge's depth to mean depth of just the edge
coordinates. If deeper than the mean, that edge is the bottom, and if
shallower, that edge is the top.