A priori triangular slip constraints can be set for a single component.
Now a single component (strike-slip or dip/tensile-slip) of triangular
slip can be set for specified elements, as opposed to setting both
components. The structure of the constraint text file remains the
same, with each line giving the element index, strike-parallel slip
rate, strike-normal slip rate. If one of these slip rates is set to
NaN, then it will be ignored, enabling constraints on a single
component of slip.